+++ src/tests/sqlsmith/src/lib.rs
@@ -26,8 +26,10 @@ use risingwave_sqlparser::parser::Parser;
 
 use crate::sql_gen::SqlGenerator;
 
+pub mod reducer;
 pub mod runner;
 mod sql_gen;
+mod utils;
 pub mod validation;
 pub use validation::is_permissible_error;
