+++ src/tests/sqlsmith/src/utils.rs
@@ -1,0 +1,29 @@
+use std::fs::File;
+use std::io::Write;
+use std::path::Path;
+
+use anyhow::{anyhow, Result};
+
+pub(crate) fn read_file_contents<P: AsRef<Path>>(filepath: P) -> Result<String> {
+    std::fs::read_to_string(filepath.as_ref()).map_err(|e| {
+        anyhow!(
+            "Failed to read contents from {} due to {e}",
+            filepath.as_ref().display()
+        )
+    })
+}
+
+pub(crate) fn create_file<P: AsRef<Path>>(filepath: P) -> Result<File> {
+    std::fs::File::create(filepath.as_ref()).map_err(|e| {
+        anyhow!(
+            "Failed to create file: {} due to {e}",
+            filepath.as_ref().display()
+        )
+    })
+}
+
+pub(crate) fn write_to_file<S: AsRef<str>>(file: &mut File, contents: S) -> Result<()> {
+    let s = contents.as_ref().as_bytes();
+    file.write_all(s)
+        .map_err(|e| anyhow!("Failed to write file due to {e}"))
+}
